crm/requirements.txt
# hello54_crm/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
psycopg2-binary==2.9.9